moved DateTimeIndicator to package indicator/helpers
