+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/DateTimeIndicator.java
@@ -21,13 +21,14 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package org.ta4j.core.indicators;
+package org.ta4j.core.indicators.helpers;
 
 import java.time.ZonedDateTime;
 import java.util.function.Function;
 
 import org.ta4j.core.Bar;
 import org.ta4j.core.BarSeries;
+import org.ta4j.core.indicators.CachedIndicator;
 
 /**
  * DateTime indicator.
